new settings
updating classifier settings to remove attributes and share the same fitBinaryClassifier function.  The settings need to contain the fitFunction that takes data and hyperparams and outputs the model/prediction (and performance?), the predictFunction that takes the model/data and returns prediction and the varImpFunction that takes the model and outputs the variable importance.